baaz/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
groq==0.9.0
httpx==0.27.0
pydantic==2.7.0
python-dotenv==1.0.1
aiosqlite==0.20.0
